VIX-3044 Organize xModel import types
Custom, circle, and state-specific xModel import code lived in the
shared xLights import namespace, making the growing importer harder
to scan as more model types are added.

Move model-specific files into Custom, Circle, and States namespaces
while keeping shared importer infrastructure in the root xLights
namespace.
