+++ src/Vixen.Modules/App/CustomPropEditor/Import/XLights/Circle/CircleXModelElementParser.cs
@@ -1,7 +1,8 @@
 using System.Xml.Linq;
 using NLog;
+using VixenModules.App.CustomPropEditor.Import.XLights;
 
-namespace VixenModules.App.CustomPropEditor.Import.XLights
+namespace VixenModules.App.CustomPropEditor.Import.XLights.Circle
 {
 	internal sealed class CircleXModelElementParser(Func<string, string, Task> showModelErrorAsync) : IXModelElementParser
 	{
