+++ src/Vixen.Modules/App/CustomPropEditor/Import/XLights/Custom/CustomXModelSource.cs
@@ -1,4 +1,4 @@
-namespace VixenModules.App.CustomPropEditor.Import.XLights
+namespace VixenModules.App.CustomPropEditor.Import.XLights.Custom
 {
 	internal enum CustomXModelSource
 	{
